// job_board_frontend_project/interfaces/Jobseeker.ts
// interfaces/Jobseeker.ts

export interface JobseekerForm {
  username: string;
  email: string;
  first_name: string;
  last_name: string;
  password: string;
  gender: string;
  ethnicity: string;
  country: string;
  state: string;
  cell_number: string;
  skills: string;
  experience: string;
  bio: string;
}